delete item auth

+++ backend/src/resolvers/Mutation.js
@@ -47,9 +47,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`)
+    const item = await ctx.db.query.item({ where }, `{id title user {id}}`)
     // 2. check if they own that item, or have permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasAuthorization = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    )
+
+    if (!ownsItem && !hasAuthorization) {
+      throw new Error('Unauthroized!')
+    }
     // 3. delete item!
     return ctx.db.mutation.deleteItem({ where }, info)
   },
